clean: split into files
