+++ .gitignore
@@ -1,1 +1,2 @@
 dataset/
+__pycache__/
