+++ src/qaset.py
@@ -1,0 +1,9 @@
+
+from dataclasses import dataclass
+
+
+@dataclass
+class QASet:
+    context: str
+    question: str
+    answer: str
